+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/integration/ModernFixCompat.java
@@ -1,0 +1,9 @@
+package net.mehvahdjukaar.moonlight.api.integration;
+
+import org.embeddedt.modernfix.core.ModernFixMixinPlugin;
+
+public class ModernFixCompat {
+    public static boolean areLazyResourcesOn(){
+        return ModernFixMixinPlugin.instance.isOptionEnabled("perf.dynamic_resources.SomeDummyClassNameHere");
+    }
+}