+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/resources/pack/DynamicTexturePack.java
@@ -15,6 +15,8 @@ import java.nio.file.Path;
 
 public class DynamicTexturePack extends DynamicResourcePack {
 
+    boolean addJsonsToStatic;
+
     public DynamicTexturePack(ResourceLocation name, Pack.Position position, boolean fixed, boolean hidden) {
         super(name, PackType.CLIENT_RESOURCES, position, fixed, hidden);
     }
@@ -64,5 +66,13 @@ public class DynamicTexturePack extends DynamicResourcePack {
 
     public void addLang(ResourceLocation langName, LangBuilder builder) {
         this.addJson(langName, builder.build(), ResType.LANG);
+    }
+
+    @Override
+    public void addJson(ResourceLocation location, JsonElement json, ResType resType) {
+        super.addJson(location, json, resType);
+        if(addJsonsToStatic){
+            persistentResources.add(location);
+        }
     }
 }